+++ src/main/java/me/gravityio/viewboboptions/VanillaOptions.java
@@ -9,32 +9,46 @@ public class VanillaOptions {
 
     public static final OptionInstance<Integer> ALL_BOBBING_STRENGTH = new OptionInstance<>(
             "options.vanilla.viewboboptions.all_bobbing_strength.label",
-            value -> Tooltip.create(Component.translatable("options.vanilla.viewboboptions.all_bobbing_strength.description")),
+            value -> Tooltip.create(
+                    Component.translatable("options.vanilla.viewboboptions.all_bobbing_strength.description")),
             (optionText, value) -> Options.genericValueLabel(optionText, Component.literal("%s%%".formatted(value))),
             new OptionInstance.IntRange(0, 100), 100, value -> {
-                ModConfig.INSTANCE.all_bobbing_strength = value.shortValue();
-                ModConfig.GSON.save();
-            }
+        ModConfig.INSTANCE.all_bobbing_strength = value.shortValue();
+        ModConfig.GSON.save();
+    }
     );
 
     public static final OptionInstance<Integer> HAND_BOBBING_STRENGTH = new OptionInstance<>(
             "options.vanilla.viewboboptions.hand_bobbing_strength.label",
-            value -> Tooltip.create(Component.translatable("options.vanilla.viewboboptions.hand_bobbing_strength.description")),
+            value -> Tooltip.create(
+                    Component.translatable("options.vanilla.viewboboptions.hand_bobbing_strength.description")),
             (optionText, value) -> Options.genericValueLabel(optionText, Component.literal("%s%%".formatted(value))),
             new OptionInstance.IntRange(0, 100), 100, value -> {
-                ModConfig.INSTANCE.hand_bobbing_strength = value.shortValue();
-                ModConfig.GSON.save();
-            }
+        ModConfig.INSTANCE.hand_bobbing_strength = value.shortValue();
+        ModConfig.GSON.save();
+    }
     );
 
     public static final OptionInstance<Integer> CAMERA_BOBBING_STRENGTH = new OptionInstance<>(
             "options.vanilla.viewboboptions.camera_bobbing_strength.label",
-            value -> Tooltip.create(Component.translatable("options.vanilla.viewboboptions.camera_bobbing_strength.description")),
+            value -> Tooltip.create(
+                    Component.translatable("options.vanilla.viewboboptions.camera_bobbing_strength.description")),
             (optionText, value) -> Options.genericValueLabel(optionText, Component.literal("%s%%".formatted(value))),
             new OptionInstance.IntRange(0, 100), 100, value -> {
-                ModConfig.INSTANCE.camera_bobbing_strength = value.shortValue();
-                ModConfig.GSON.save();
-            }
+        ModConfig.INSTANCE.camera_bobbing_strength = value.shortValue();
+        ModConfig.GSON.save();
+    }
+    );
+
+    public static final OptionInstance<Integer> HAND_SWAY_STRENGTH = new OptionInstance<>(
+            "options.vanilla.viewboboptions.hand_sway_strength.label",
+            value -> Tooltip.create(
+                    Component.translatable("options.vanilla.viewboboptions.hand_sway_strength.description")),
+            (optionText, value) -> Options.genericValueLabel(optionText, Component.literal("%s%%".formatted(value))),
+            new OptionInstance.IntRange(0, 100), 100, value -> {
+        ModConfig.INSTANCE.hand_sway_strength = value.shortValue();
+        ModConfig.GSON.save();
+    }
     );
 
 
@@ -42,5 +56,6 @@ public class VanillaOptions {
         ALL_BOBBING_STRENGTH.set((int) ModConfig.INSTANCE.all_bobbing_strength);
         HAND_BOBBING_STRENGTH.set((int) ModConfig.INSTANCE.hand_bobbing_strength);
         CAMERA_BOBBING_STRENGTH.set((int) ModConfig.INSTANCE.camera_bobbing_strength);
+        HAND_SWAY_STRENGTH.set((int) ModConfig.INSTANCE.hand_sway_strength);
     }
 }